requirements.txt
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
amqp==5.2.0
anyio==4.4.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==2.15.8
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
black==23.9.1
bleach==6.1.0
boto3==1.35.19
botocore==1.35.19
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
constantly==23.10.4
CounterFit==0.1.4.dev9
coverage==7.6.1
crispy-bootstrap5==0.7
cron-descriptor==1.4.5
cryptography==42.0.8
cssbeautifier==1.15.1
cssselect2==0.7.0
cvlib==0.2.7
Cython==3.0.9
daphne==4.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.16
django-appconf==1.0.6
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.3.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-select2==8.1.2
django-timezone-field==7.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djlint==1.34.0
dnspython==2.6.1
docutils==0.21.2
EditorConfig==0.12.4
et-xmlfile==1.1.0
eventlet==0.33.3
executing==2.1.0
factory-boy==3.3.0
Faker==28.4.1
fastjsonschema==2.20.0
filelock==3.16.0
flake8==7.1.1
flake8-isort==6.1.0
Flask==2.1.2
Flask-SocketIO==5.2.0
fqdn==1.5.1
frozenlist==1.4.1
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperlink==21.0.0
identify==2.6.1
idna==3.9
imageio==2.35.1
imagesize==1.4.1
imutils==0.5.4
incremental==24.7.2
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kombu==5.4.1
lazy-object-proxy==1.10.0
livereload==2.7.0
loguru==0.7.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
msgpack==1.1.0
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.1
openpyxl==3.1.5
oscrypto==1.3.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.11.2
pexpect==4.9.0
phonenumbers==8.13.45
pillow==10.4.0
pip-check-reqs==2.5.3
pip-chill==1.0.3
platformdirs==4.3.3
pluggy==1.5.0
pre-commit==3.4.0
progressbar==2.5
# psycopg2
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyOpenSSL==24.0.0
pypdf==4.3.1
pypng==0.20220715.0
pysafebrowsing==0.1.2
pytest==8.3.3
pytest-django==4.5.2
pytest-sugar==0.9.7
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.9.1
python-json-logger==2.0.7
python-slugify==8.0.1
python-socketio==5.11.4
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
qtconsole==5.6.0
QtPy==2.4.1
redis==5.0.8
# regex==2023.12.25
referencing==0.35.1
regex==2023.12.25
reportlab==4.0.9
requests==2.32.3
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
s3transfer==0.10.2
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==74.1.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
svglib==1.5.1
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
twilio==9.3.0
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240906
types-pytz==2024.2.0.20240913
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.23.2
vine==5.1.0
virtualenv==20.26.4
waitress==3.0.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
whitenoise==6.7.0
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.15
yarl==1.11.1
zope.interface==6.1
psycopg2-binary
pytz
django_browser_reload